+++ app/_pages/CreatePage/PromptInput/StylePresetsDropdown/presets.ts
@@ -68,10 +68,22 @@ export const stableHordeStyles: { [key: string]: HordePreset } = {
     hires_fix: true,
     cfg_scale: 9,
     loras: [
-      { name: '85479', model: 0.5 },
-      { name: '42190', model: 0.85 },
-      { name: '109775', model: 0.2 },
-      { name: '22437', model: 0.85 }
+      {
+        name: '85479',
+        model: 0.5
+      },
+      {
+        name: '42190',
+        model: 0.85
+      },
+      {
+        name: '109775',
+        model: 0.2
+      },
+      {
+        name: '22437',
+        model: 0.85
+      }
     ]
   },
   'sdxl-vertical': {
@@ -130,7 +142,48 @@ export const stableHordeStyles: { [key: string]: HordePreset } = {
     height: 640,
     steps: 50
   },
-
+  'hamunaptra party': {
+    prompt: '{p}###{np}',
+    model: 'Dreamshaper',
+    sampler_name: 'k_dpmpp_2m',
+    steps: 27,
+    cfg_scale: 7.5,
+    width: 768,
+    height: 960,
+    hires_fix: true,
+    tis: [
+      {
+        name: '4157'
+      }
+    ]
+  },
+  bootybash: {
+    prompt:
+      '{p}###{np}, low quality, blurry, no pupils, missing digits, extra limbs, monochrome, watermark, signature, artist name, bad face, bad hand, worst quality, writing,',
+    model: 'Anything Diffusion',
+    sampler: 'k_euler',
+    karras: true,
+    steps: 30,
+    cfg_scale: 7,
+    width: 768,
+    height: 1024,
+    clip_skip: true,
+    hires_fix: true,
+    loras: [
+      {
+        name: '58390',
+        model: 0.55,
+        clip: 1
+      }
+    ],
+    tis: [
+      {
+        name: '7808',
+        inject_ti: 'negprompt',
+        strength: 0.7
+      }
+    ]
+  },
   graffiti: {
     prompt:
       'graffiti style {p}, street art, vibrant, urban, detailed, tag, mural###{np}, ugly, deformed, noisy, blurry, low contrast, realism, photorealistic',
@@ -269,7 +322,7 @@ export const stableHordeStyles: { [key: string]: HordePreset } = {
   },
   goldenagecomics: {
     prompt:
-      '{np}, 1940s golden age comics style screen print, cover art by CC Beck and Will Eisner, (thick lines:0.7), early CMYK lithography, halftones###3d, modern, black and white, spot color{np}',
+      '{p}, 1940s golden age comics style screen print, cover art by CC Beck and Will Eisner, (thick lines:0.7), early CMYK lithography, halftones###3d, modern, black and white, spot color{np}',
     model: 'SDXL 1.0',
     width: 832,
     height: 1280,
@@ -533,7 +586,6 @@ export const stableHordeStyles: { [key: string]: HordePreset } = {
     steps: 50,
     cfg_scale: 7
   },
-
   glitchart: {
     prompt:
       '{p}###{np}, worst quality, low quality:1.4), EasyNegative, bad anatomy, bad hands, cropped, missing fingers, missing toes, too many toes, too many fingers, missing arms, long neck, Humpbacked, deformed, disfigured, poorly drawn face, distorted face, mutation, mutated, extra limb, ugly, poorly drawn hands, missing limb, floating limbs, disconnected limbs, malformed hands, out of focus, long body, monochrome, symbol, text, logo, door frame, window frame, mirror frame',
@@ -542,7 +594,12 @@ export const stableHordeStyles: { [key: string]: HordePreset } = {
     cfg_scale: 7.5,
     width: 512,
     height: 640,
-    loras: [{ name: '71125', inject_trigger: 'any' }]
+    loras: [
+      {
+        name: '71125',
+        inject_trigger: 'any'
+      }
+    ]
   },
   oldvhs: {
     prompt:
@@ -552,7 +609,11 @@ export const stableHordeStyles: { [key: string]: HordePreset } = {
     cfg_scale: 7.5,
     width: 512,
     height: 640,
-    loras: [{ name: '103667' }]
+    loras: [
+      {
+        name: '103667'
+      }
+    ]
   },
   pixelsorting: {
     prompt:
@@ -562,7 +623,12 @@ export const stableHordeStyles: { [key: string]: HordePreset } = {
     cfg_scale: 7.5,
     width: 512,
     height: 640,
-    loras: [{ name: '57963', inject_trigger: 'any' }]
+    loras: [
+      {
+        name: '57963',
+        inject_trigger: 'any'
+      }
+    ]
   },
   analogdistortion: {
     prompt:
@@ -572,16 +638,24 @@ export const stableHordeStyles: { [key: string]: HordePreset } = {
     cfg_scale: 7.5,
     width: 512,
     height: 640,
-    loras: [{ name: '104917' }]
+    loras: [
+      {
+        name: '104917'
+      }
+    ]
   },
-
   gothichorrorai: {
     prompt:
       '{p}###{np}, worst quality, low quality:1.4), EasyNegative, bad anatomy, bad hands, cropped, missing fingers, missing toes, too many toes, too many fingers, missing arms, long neck, Humpbacked, deformed, disfigured, poorly drawn face, distorted face, mutation, mutated, extra limb, ugly, poorly drawn hands, missing limb, floating limbs, disconnected limbs, malformed hands, out of focus, long body, monochrome, symbol, text, logo, door frame, window frame, mirror frame',
     model: 'Dreamshaper',
     steps: 40,
     cfg_scale: 7.5,
-    loras: [{ name: '39760', inject_trigger: 'any' }]
+    loras: [
+      {
+        name: '39760',
+        inject_trigger: 'any'
+      }
+    ]
   },
   tentaclehorrorai: {
     prompt:
@@ -589,7 +663,12 @@ export const stableHordeStyles: { [key: string]: HordePreset } = {
     model: 'Dreamshaper',
     steps: 40,
     cfg_scale: 7.5,
-    loras: [{ name: '42468', inject_trigger: 'any' }]
+    loras: [
+      {
+        name: '42468',
+        inject_trigger: 'any'
+      }
+    ]
   },
   reelhorror: {
     prompt:
@@ -597,7 +676,11 @@ export const stableHordeStyles: { [key: string]: HordePreset } = {
     model: 'Dreamshaper',
     steps: 40,
     cfg_scale: 7.5,
-    loras: [{ name: '67666' }]
+    loras: [
+      {
+        name: '67666'
+      }
+    ]
   },
   bodyhorror: {
     prompt:
@@ -627,7 +710,11 @@ export const stableHordeStyles: { [key: string]: HordePreset } = {
         clip: 1,
         inject_trigger: 'body horror, mutation, blood'
       },
-      { name: '67666', model: 0.7, clip: 1 }
+      {
+        name: '67666',
+        model: 0.7,
+        clip: 1
+      }
     ]
   },
   ivorygoldai: {
@@ -636,7 +723,12 @@ export const stableHordeStyles: { [key: string]: HordePreset } = {
     model: 'Dreamshaper',
     steps: 40,
     cfg_scale: 7.5,
-    loras: [{ name: '62700', inject_trigger: 'any' }]
+    loras: [
+      {
+        name: '62700',
+        inject_trigger: 'any'
+      }
+    ]
   },
   arielphorror: {
     prompt:
@@ -644,7 +736,11 @@ export const stableHordeStyles: { [key: string]: HordePreset } = {
     model: 'Lyriel',
     steps: 40,
     cfg_scale: 7.5,
-    loras: [{ name: '39760' }]
+    loras: [
+      {
+        name: '39760'
+      }
+    ]
   },
   dreadhorror: {
     prompt:
@@ -652,9 +748,12 @@ export const stableHordeStyles: { [key: string]: HordePreset } = {
     model: 'Deliberate',
     steps: 40,
     cfg_scale: 7.5,
-    loras: [{ name: '98011' }]
+    loras: [
+      {
+        name: '98011'
+      }
+    ]
   },
-
   'vibrant halloween': {
     prompt:
       'a Halloween themed vibrant pulp art by JC leyendecker and norman rockwell featuring {p}. horror vibe, dark atmospthere, depth of field, campy, spooky. purple, orange and green color palette.###{np}',
@@ -729,14 +828,17 @@ export const stableHordeStyles: { [key: string]: HordePreset } = {
     height: 1024,
     cfg_scale: 9
   },
-
   piratepunk: {
     prompt:
       '{p}, piratepunkai###{np}, worst quality, low quality:1.4), EasyNegative, bad anatomy, bad hands, cropped, missing fingers, missing toes, too many toes, too many fingers, missing arms, long neck, Humpbacked, deformed, disfigured, poorly drawn face, distorted face, mutation, mutated, extra limb, ugly, poorly drawn hands, missing limb, floating limbs, disconnected limbs, malformed hands, out of focus, long body, monochrome, symbol, text, logo, door frame, window frame, mirror frame',
     model: 'Dreamshaper',
     steps: 40,
     cfg_scale: 7.5,
-    loras: [{ name: '45892' }]
+    loras: [
+      {
+        name: '45892'
+      }
+    ]
   },
   lunarpunk: {
     prompt:
@@ -744,7 +846,12 @@ export const stableHordeStyles: { [key: string]: HordePreset } = {
     model: 'Dreamshaper',
     steps: 40,
     cfg_scale: 7.5,
-    loras: [{ name: '97136', inject_trigger: 'any' }]
+    loras: [
+      {
+        name: '97136',
+        inject_trigger: 'any'
+      }
+    ]
   },
   celtpunk: {
     prompt:
@@ -752,7 +859,12 @@ export const stableHordeStyles: { [key: string]: HordePreset } = {
     model: 'Dreamshaper',
     steps: 40,
     cfg_scale: 7.5,
-    loras: [{ name: '92008', inject_trigger: 'any' }]
+    loras: [
+      {
+        name: '92008',
+        inject_trigger: 'any'
+      }
+    ]
   },
   toxicpunk: {
     prompt:
@@ -760,7 +872,12 @@ export const stableHordeStyles: { [key: string]: HordePreset } = {
     model: 'Dreamshaper',
     steps: 40,
     cfg_scale: 7.5,
-    loras: [{ name: '95802', inject_trigger: 'any' }]
+    loras: [
+      {
+        name: '95802',
+        inject_trigger: 'any'
+      }
+    ]
   },
   mongolpunk: {
     prompt:
@@ -768,7 +885,12 @@ export const stableHordeStyles: { [key: string]: HordePreset } = {
     model: 'Dreamshaper',
     steps: 40,
     cfg_scale: 7.5,
-    loras: [{ name: '85109', inject_trigger: 'any' }]
+    loras: [
+      {
+        name: '85109',
+        inject_trigger: 'any'
+      }
+    ]
   },
   musketpunk: {
     prompt:
@@ -776,7 +898,12 @@ export const stableHordeStyles: { [key: string]: HordePreset } = {
     model: 'Dreamshaper',
     steps: 40,
     cfg_scale: 7.5,
-    loras: [{ name: '90543', inject_trigger: 'any' }]
+    loras: [
+      {
+        name: '90543',
+        inject_trigger: 'any'
+      }
+    ]
   },
   absinthepunk: {
     prompt:
@@ -784,7 +911,12 @@ export const stableHordeStyles: { [key: string]: HordePreset } = {
     model: 'Dreamshaper',
     steps: 40,
     cfg_scale: 7.5,
-    loras: [{ name: '89183', inject_trigger: 'any' }]
+    loras: [
+      {
+        name: '89183',
+        inject_trigger: 'any'
+      }
+    ]
   },
   cattlepunk: {
     prompt:
@@ -792,7 +924,12 @@ export const stableHordeStyles: { [key: string]: HordePreset } = {
     model: 'Dreamshaper',
     steps: 40,
     cfg_scale: 7.5,
-    loras: [{ name: '95828', inject_trigger: 'any' }]
+    loras: [
+      {
+        name: '95828',
+        inject_trigger: 'any'
+      }
+    ]
   },
   circuspunk: {
     prompt:
@@ -800,7 +937,12 @@ export const stableHordeStyles: { [key: string]: HordePreset } = {
     model: 'Dreamshaper',
     steps: 40,
     cfg_scale: 7.5,
-    loras: [{ name: '87884', inject_trigger: 'any' }]
+    loras: [
+      {
+        name: '87884',
+        inject_trigger: 'any'
+      }
+    ]
   },
   aetherpunk: {
     prompt:
@@ -808,7 +950,12 @@ export const stableHordeStyles: { [key: string]: HordePreset } = {
     model: 'Dreamshaper',
     steps: 40,
     cfg_scale: 7.5,
-    loras: [{ name: '85036', inject_trigger: 'any' }]
+    loras: [
+      {
+        name: '85036',
+        inject_trigger: 'any'
+      }
+    ]
   },
   potatopunk: {
     prompt:
@@ -816,7 +963,12 @@ export const stableHordeStyles: { [key: string]: HordePreset } = {
     model: 'Dreamshaper',
     steps: 40,
     cfg_scale: 7.5,
-    loras: [{ name: '72231', inject_trigger: 'any' }]
+    loras: [
+      {
+        name: '72231',
+        inject_trigger: 'any'
+      }
+    ]
   },
   coalpunk: {
     prompt:
@@ -824,7 +976,11 @@ export const stableHordeStyles: { [key: string]: HordePreset } = {
     model: 'Dreamshaper',
     steps: 40,
     cfg_scale: 7.5,
-    loras: [{ name: '84753' }]
+    loras: [
+      {
+        name: '84753'
+      }
+    ]
   },
   alchemypunk: {
     prompt:
@@ -832,7 +988,11 @@ export const stableHordeStyles: { [key: string]: HordePreset } = {
     model: 'Dreamshaper',
     steps: 40,
     cfg_scale: 7.5,
-    loras: [{ name: '78766' }]
+    loras: [
+      {
+        name: '78766'
+      }
+    ]
   },
   gothicpunk: {
     prompt:
@@ -840,7 +1000,11 @@ export const stableHordeStyles: { [key: string]: HordePreset } = {
     model: 'Dreamshaper',
     steps: 40,
     cfg_scale: 7.5,
-    loras: [{ name: '78695' }]
+    loras: [
+      {
+        name: '78695'
+      }
+    ]
   },
   witchpunk: {
     prompt:
@@ -848,7 +1012,11 @@ export const stableHordeStyles: { [key: string]: HordePreset } = {
     model: 'Dreamshaper',
     steps: 40,
     cfg_scale: 7.5,
-    loras: [{ name: '78280' }]
+    loras: [
+      {
+        name: '78280'
+      }
+    ]
   },
   colorpunk: {
     prompt:
@@ -856,7 +1024,11 @@ export const stableHordeStyles: { [key: string]: HordePreset } = {
     model: 'Dreamshaper',
     steps: 40,
     cfg_scale: 7.5,
-    loras: [{ name: '78060' }]
+    loras: [
+      {
+        name: '78060'
+      }
+    ]
   },
   bronzepunk: {
     prompt:
@@ -864,7 +1036,11 @@ export const stableHordeStyles: { [key: string]: HordePreset } = {
     model: 'Dreamshaper',
     steps: 40,
     cfg_scale: 7.5,
-    loras: [{ name: '77293' }]
+    loras: [
+      {
+        name: '77293'
+      }
+    ]
   },
   neonpunk: {
     prompt:
@@ -872,7 +1048,11 @@ export const stableHordeStyles: { [key: string]: HordePreset } = {
     model: 'Dreamshaper',
     steps: 40,
     cfg_scale: 7.5,
-    loras: [{ name: '77121' }]
+    loras: [
+      {
+        name: '77121'
+      }
+    ]
   },
   kazakpunk: {
     prompt:
@@ -880,7 +1060,11 @@ export const stableHordeStyles: { [key: string]: HordePreset } = {
     model: 'Dreamshaper',
     steps: 40,
     cfg_scale: 7.5,
-    loras: [{ name: '76737' }]
+    loras: [
+      {
+        name: '76737'
+      }
+    ]
   },
   grecopunk: {
     prompt:
@@ -888,7 +1072,11 @@ export const stableHordeStyles: { [key: string]: HordePreset } = {
     model: 'Dreamshaper',
     steps: 40,
     cfg_scale: 7.5,
-    loras: [{ name: '76190' }]
+    loras: [
+      {
+        name: '76190'
+      }
+    ]
   },
   romanpunk: {
     prompt:
@@ -896,7 +1084,11 @@ export const stableHordeStyles: { [key: string]: HordePreset } = {
     model: 'Dreamshaper',
     steps: 40,
     cfg_scale: 7.5,
-    loras: [{ name: '76190' }]
+    loras: [
+      {
+        name: '76190'
+      }
+    ]
   },
   sovietpunk: {
     prompt:
@@ -904,7 +1096,11 @@ export const stableHordeStyles: { [key: string]: HordePreset } = {
     model: 'Dreamshaper',
     steps: 40,
     cfg_scale: 7.5,
-    loras: [{ name: '75709' }]
+    loras: [
+      {
+        name: '75709'
+      }
+    ]
   },
   arabiannightspunk: {
     prompt:
@@ -912,7 +1108,11 @@ export const stableHordeStyles: { [key: string]: HordePreset } = {
     model: 'Dreamshaper',
     steps: 40,
     cfg_scale: 7.5,
-    loras: [{ name: '69315' }]
+    loras: [
+      {
+        name: '69315'
+      }
+    ]
   },
   steampunk: {
     prompt:
@@ -920,7 +1120,11 @@ export const stableHordeStyles: { [key: string]: HordePreset } = {
     model: 'Dreamshaper',
     steps: 40,
     cfg_scale: 7.5,
-    loras: [{ name: '59338' }]
+    loras: [
+      {
+        name: '59338'
+      }
+    ]
   },
   cogpunk: {
     prompt:
@@ -928,7 +1132,11 @@ export const stableHordeStyles: { [key: string]: HordePreset } = {
     model: 'Dreamshaper',
     steps: 40,
     cfg_scale: 7.5,
-    loras: [{ name: '59338' }]
+    loras: [
+      {
+        name: '59338'
+      }
+    ]
   },
   victorianpunk: {
     prompt:
@@ -936,7 +1144,11 @@ export const stableHordeStyles: { [key: string]: HordePreset } = {
     model: 'Dreamshaper',
     steps: 40,
     cfg_scale: 7.5,
-    loras: [{ name: '84250' }]
+    loras: [
+      {
+        name: '84250'
+      }
+    ]
   },
   biopunk: {
     prompt:
@@ -944,7 +1156,11 @@ export const stableHordeStyles: { [key: string]: HordePreset } = {
     model: 'Dreamshaper',
     steps: 40,
     cfg_scale: 7.5,
-    loras: [{ name: '80375' }]
+    loras: [
+      {
+        name: '80375'
+      }
+    ]
   },
   mahabharatapunk: {
     prompt:
@@ -952,7 +1168,11 @@ export const stableHordeStyles: { [key: string]: HordePreset } = {
     model: 'Dreamshaper',
     steps: 40,
     cfg_scale: 7.5,
-    loras: [{ name: '71478' }]
+    loras: [
+      {
+        name: '71478'
+      }
+    ]
   },
   javapunk: {
     prompt:
@@ -960,7 +1180,11 @@ export const stableHordeStyles: { [key: string]: HordePreset } = {
     model: 'Dreamshaper',
     steps: 40,
     cfg_scale: 7.5,
-    loras: [{ name: '72415' }]
+    loras: [
+      {
+        name: '72415'
+      }
+    ]
   },
   balipunk: {
     prompt:
@@ -968,7 +1192,11 @@ export const stableHordeStyles: { [key: string]: HordePreset } = {
     model: 'Dreamshaper',
     steps: 40,
     cfg_scale: 7.5,
-    loras: [{ name: '72347' }]
+    loras: [
+      {
+        name: '72347'
+      }
+    ]
   },
   dieselpunk: {
     prompt:
@@ -976,7 +1204,11 @@ export const stableHordeStyles: { [key: string]: HordePreset } = {
     model: 'Dreamshaper',
     steps: 40,
     cfg_scale: 7.5,
-    loras: [{ name: '64676' }]
+    loras: [
+      {
+        name: '64676'
+      }
+    ]
   },
   totempunk: {
     prompt:
@@ -984,7 +1216,11 @@ export const stableHordeStyles: { [key: string]: HordePreset } = {
     model: 'Dreamshaper',
     steps: 40,
     cfg_scale: 7.5,
-    loras: [{ name: '71819' }]
+    loras: [
+      {
+        name: '71819'
+      }
+    ]
   },
   samuraipunk: {
     prompt:
@@ -992,7 +1228,11 @@ export const stableHordeStyles: { [key: string]: HordePreset } = {
     model: 'Dreamshaper',
     steps: 40,
     cfg_scale: 7.5,
-    loras: [{ name: '69108' }]
+    loras: [
+      {
+        name: '69108'
+      }
+    ]
   },
   zulupunk: {
     prompt:
@@ -1000,7 +1240,11 @@ export const stableHordeStyles: { [key: string]: HordePreset } = {
     model: 'Dreamshaper',
     steps: 40,
     cfg_scale: 7.5,
-    loras: [{ name: '65725' }]
+    loras: [
+      {
+        name: '65725'
+      }
+    ]
   },
   teslapunk: {
     prompt:
@@ -1008,7 +1252,11 @@ export const stableHordeStyles: { [key: string]: HordePreset } = {
     model: 'Dreamshaper',
     steps: 40,
     cfg_scale: 7.5,
-    loras: [{ name: '64955' }]
+    loras: [
+      {
+        name: '64955'
+      }
+    ]
   },
   valvepunk: {
     prompt:
@@ -1016,7 +1264,11 @@ export const stableHordeStyles: { [key: string]: HordePreset } = {
     model: 'Dreamshaper',
     steps: 40,
     cfg_scale: 7.5,
-    loras: [{ name: '64955' }]
+    loras: [
+      {
+        name: '64955'
+      }
+    ]
   },
   infernalpunk: {
     prompt:
@@ -1024,7 +1276,11 @@ export const stableHordeStyles: { [key: string]: HordePreset } = {
     model: 'Dreamshaper',
     steps: 40,
     cfg_scale: 7.5,
-    loras: [{ name: '62585' }]
+    loras: [
+      {
+        name: '62585'
+      }
+    ]
   },
   solarpunk: {
     prompt:
@@ -1032,7 +1288,11 @@ export const stableHordeStyles: { [key: string]: HordePreset } = {
     model: 'Dreamshaper',
     steps: 40,
     cfg_scale: 7.5,
-    loras: [{ name: '43944' }]
+    loras: [
+      {
+        name: '43944'
+      }
+    ]
   },
   stonepunk: {
     prompt:
@@ -1040,7 +1300,11 @@ export const stableHordeStyles: { [key: string]: HordePreset } = {
     model: 'Dreamshaper',
     steps: 40,
     cfg_scale: 7.5,
-    loras: [{ name: '51539' }]
+    loras: [
+      {
+        name: '51539'
+      }
+    ]
   },
   gempunk: {
     prompt:
@@ -1048,7 +1312,11 @@ export const stableHordeStyles: { [key: string]: HordePreset } = {
     model: 'Dreamshaper',
     steps: 40,
     cfg_scale: 7.5,
-    loras: [{ name: '49374' }]
+    loras: [
+      {
+        name: '49374'
+      }
+    ]
   },
   horrorpunk: {
     prompt:
@@ -1056,7 +1324,11 @@ export const stableHordeStyles: { [key: string]: HordePreset } = {
     model: 'Dreamshaper',
     steps: 40,
     cfg_scale: 7.5,
-    loras: [{ name: '47489' }]
+    loras: [
+      {
+        name: '47489'
+      }
+    ]
   },
   bonepunk: {
     prompt:
@@ -1064,7 +1336,11 @@ export const stableHordeStyles: { [key: string]: HordePreset } = {
     model: 'Dreamshaper',
     steps: 40,
     cfg_scale: 7.5,
-    loras: [{ name: '48356' }]
+    loras: [
+      {
+        name: '48356'
+      }
+    ]
   },
   oldegyptpunk: {
     prompt:
@@ -1072,7 +1348,11 @@ export const stableHordeStyles: { [key: string]: HordePreset } = {
     model: 'Dreamshaper',
     steps: 40,
     cfg_scale: 7.5,
-    loras: [{ name: '43229' }]
+    loras: [
+      {
+        name: '43229'
+      }
+    ]
   },
   vikingpunk: {
     prompt:
@@ -1080,7 +1360,11 @@ export const stableHordeStyles: { [key: string]: HordePreset } = {
     model: 'Dreamshaper',
     steps: 40,
     cfg_scale: 7.5,
-    loras: [{ name: '41364' }]
+    loras: [
+      {
+        name: '41364'
+      }
+    ]
   },
   gaslamppunk: {
     prompt:
@@ -1088,7 +1372,11 @@ export const stableHordeStyles: { [key: string]: HordePreset } = {
     model: 'Dreamshaper',
     steps: 40,
     cfg_scale: 7.5,
-    loras: [{ name: '71335' }]
+    loras: [
+      {
+        name: '71335'
+      }
+    ]
   },
   atompunk: {
     prompt:
@@ -1096,7 +1384,11 @@ export const stableHordeStyles: { [key: string]: HordePreset } = {
     model: 'Dreamshaper',
     steps: 40,
     cfg_scale: 7.5,
-    loras: [{ name: '64235' }]
+    loras: [
+      {
+        name: '64235'
+      }
+    ]
   },
   gothpunk: {
     prompt:
@@ -1104,7 +1396,11 @@ export const stableHordeStyles: { [key: string]: HordePreset } = {
     model: 'Dreamshaper',
     steps: 40,
     cfg_scale: 7.5,
-    loras: [{ name: '75265' }]
+    loras: [
+      {
+        name: '75265'
+      }
+    ]
   },
   circuitpunk: {
     prompt:
@@ -1112,7 +1408,11 @@ export const stableHordeStyles: { [key: string]: HordePreset } = {
     model: 'Dreamshaper',
     steps: 40,
     cfg_scale: 7.5,
-    loras: [{ name: '58410' }]
+    loras: [
+      {
+        name: '58410'
+      }
+    ]
   },
   nightmarepunk: {
     prompt:
@@ -1120,7 +1420,11 @@ export const stableHordeStyles: { [key: string]: HordePreset } = {
     model: 'Dreamshaper',
     steps: 40,
     cfg_scale: 7.5,
-    loras: [{ name: '56336' }]
+    loras: [
+      {
+        name: '56336'
+      }
+    ]
   },
   crystalpunk: {
     prompt:
@@ -1128,7 +1432,11 @@ export const stableHordeStyles: { [key: string]: HordePreset } = {
     model: 'Dreamshaper',
     steps: 40,
     cfg_scale: 7.5,
-    loras: [{ name: '48859' }]
+    loras: [
+      {
+        name: '48859'
+      }
+    ]
   },
   fantasypunk: {
     prompt:
@@ -1136,9 +1444,12 @@ export const stableHordeStyles: { [key: string]: HordePreset } = {
     model: 'Dreamshaper',
     steps: 40,
     cfg_scale: 7.5,
-    loras: [{ name: '42260' }]
+    loras: [
+      {
+        name: '42260'
+      }
+    ]
   },
-
   'perfect summer': {
     prompt:
       'digital art, cg painting, (medium long shot:1.2), hyper detailed, volumetric lighting, godrays, realistic, photorealistic render, golden hour, sun-drenched panorama, lush tropical colors, subtle shading, {p}, beautiful face, wide glossy eyes, ultra detailed, masterpiece, high resolution illustration, 4k, centered composition###nsfw, lowres, text, error, cropped, out of frame, worst quality, low quality, normal quality, jpeg artifacts, signature, watermark, username, blurry, out of focus, censorship, ugly, black and white, monochrome, artist name, signature, multiple views, power lines{np}',
@@ -1168,7 +1479,6 @@ export const stableHordeStyles: { [key: string]: HordePreset } = {
       'behind the scenes footage from the set of {p}, movie set, breakdown, dramatic lighting, off centered focus, clear wide shot, dreamlikeart###blurry, distorted, ugly faces,{np}',
     model: 'Dreamlike Diffusion'
   },
-
   'generic scifi': {
     prompt:
       '{p}, science fiction artwork, epic, detailed and intricate digital concept art, by John Jude Palencar, Manchu, and Devin Elle Kurtz, 4K{np}',
@@ -1311,11 +1621,15 @@ export const stableHordeStyles: { [key: string]: HordePreset } = {
       'beautiful warrior fairy {p}, portrait, up close, cinematic, snowy background, dramatic lighting, highly detailed, very intricate, meta, fanciful, optimistic, wide shot, volumetric lighting, bokeh blur, unreal engine, octane render ### deformed, ugly, mutated, watermark, trademark, border, nude, NSFW{np}',
     model: 'stable_diffusion'
   },
-
   cyberrealistic: {
     prompt: '{p}###{np}',
     model: 'CyberRealistic',
-    tis: [{ name: '77976', inject_ti: 'negprompt' }]
+    tis: [
+      {
+        name: '77976',
+        inject_ti: 'negprompt'
+      }
+    ]
   },
   deliberate: {
     prompt: '{p}###{np}',
@@ -1405,12 +1719,28 @@ export const stableHordeStyles: { [key: string]: HordePreset } = {
     hires_fix: true,
     cfg_scale: 7.5,
     loras: [
-      { name: '82098', model: 0.5 },
-      { name: '48139', model: 0.2 },
-      { name: '132532', model: -0.7 },
-      { name: '58390', model: 0.5 }
+      {
+        name: '82098',
+        model: 0.5
+      },
+      {
+        name: '48139',
+        model: 0.2
+      },
+      {
+        name: '132532',
+        model: -0.7
+      },
+      {
+        name: '58390',
+        model: 0.5
+      }
     ],
-    tis: [{ name: '16993' }]
+    tis: [
+      {
+        name: '16993'
+      }
+    ]
   },
   revanimated: {
     prompt: '{p}###{np}',